requirements.txt
google-generativeai
pillow
python-dotenv
streamlit
streamlit-drawable-canvas